Refactoring paywall flow and support play trailer

+++ app/src/main/java/com/zype/android/ui/player/PlayerFragment.java
@@ -82,6 +82,7 @@ import com.zype.android.ui.chromecast.LivePlayerActivity;
 import com.zype.android.ui.dialog.ErrorDialogFragment;
 import com.zype.android.ui.dialog.SubtitlesDialogFragment;
 import com.zype.android.ui.video_details.VideoDetailActivity;
+import com.zype.android.ui.video_details.VideoDetailViewModel;
 import com.zype.android.ui.video_details.fragments.video.MediaControlInterface;
 import com.zype.android.ui.video_details.fragments.video.OnVideoAudioListener;
 import com.zype.android.utils.BundleConstants;
@@ -121,6 +122,7 @@ public class PlayerFragment extends BaseFragment implements
     public static final int TYPE_VIDEO_WEB = 4;
     public static final int TYPE_AUDIO_LIVE = 5;
     public static final int TYPE_VIDEO_LIVE = 6;
+    public static final int TYPE_VIDEO_TRAILER = 7;
 
     public static final String CONTENT_TYPE_TYPE = "content_type";
     public static final String CONTENT_URL = "content_url";
@@ -198,6 +200,7 @@ public class PlayerFragment extends BaseFragment implements
     // Sensors
     PlayerViewModel playerViewModel;
     SensorViewModel sensorViewModel;
+    VideoDetailViewModel videoDetailViewModel;
 
     public static PlayerFragment newInstance(int mediaType, String filePath, String fileId) {
         PlayerFragment fragment = new PlayerFragment();
@@ -344,6 +347,7 @@ public class PlayerFragment extends BaseFragment implements
         });
 
         sensorViewModel = ViewModelProviders.of(getActivity()).get(SensorViewModel.class);
+        videoDetailViewModel = ViewModelProviders.of(getActivity()).get(VideoDetailViewModel.class);
 
         setHasOptionsMenu(true);
     }
@@ -408,7 +412,9 @@ public class PlayerFragment extends BaseFragment implements
         mListener.onFullscreenChanged(UiUtils.isLandscapeOrientation(getActivity()));
         registerReceivers();
         if (player == null) {
-            analytics = VideoHelper.getAnalytics(getActivity().getContentResolver(), fileId);
+            if (!TextUtils.isEmpty(fileId)) {
+                analytics = VideoHelper.getAnalytics(getActivity().getContentResolver(), fileId);
+            }
             preparePlayer(true);
         }
         else {
@@ -416,7 +422,9 @@ public class PlayerFragment extends BaseFragment implements
                 player.setBackgrounded(false);
             }
 
-            analytics = VideoHelper.getAnalytics(getActivity().getContentResolver(), fileId);
+            if (!TextUtils.isEmpty(fileId)) {
+                analytics = VideoHelper.getAnalytics(getActivity().getContentResolver(), fileId);
+            }
             attachPlayerToAnalyticsManager();
 
             player.getPlayerControl().start();
@@ -686,6 +694,7 @@ public class PlayerFragment extends BaseFragment implements
             case TYPE_VIDEO_LIVE:
             case TYPE_AUDIO_WEB:
             case TYPE_AUDIO_LIVE:
+            case TYPE_VIDEO_TRAILER:
                 if (contentUri != null &&
                         (contentUri.contains(".mp4") || contentUri.contains(".m4a") || contentUri.contains(".mp3"))) {
                     return new ExtractorRendererBuilder(getContext(), userAgent, Uri.parse(contentUri), new Mp4Extractor());
@@ -793,7 +802,7 @@ public class PlayerFragment extends BaseFragment implements
         player.setSurface(surfaceView.getHolder().getSurface());
         player.setPlayWhenReady(playWhenReady);
 
-        if (playWhenReady && !adSchedule.isEmpty()) {
+        if (playWhenReady && adSchedule != null && !adSchedule.isEmpty() && !TextUtils.isEmpty(fileId)) {
             long playTime = DataHelper.getPlayTime(getActivity().getContentResolver(), fileId);
             nextAdIndex = seekAdByPosition(playTime);
             if (!checkNextAd(playTime)) {
@@ -842,7 +851,13 @@ public class PlayerFragment extends BaseFragment implements
                 showControls();
                 if (contentType == TYPE_VIDEO_LOCAL || contentType == TYPE_VIDEO_WEB) {
                     if (playWhenReady) {
-                        mListener.videoFinished();
+                        if (playerViewModel.isPlayTrailer()) {
+                            videoDetailViewModel.onVideoFinished(true);
+                            playerViewModel.setTrailerVideoId(null);
+                        }
+                        else {
+                            mListener.videoFinished();
+                        }
                     }
                 } else if (contentType == TYPE_AUDIO_LOCAL || contentType == TYPE_AUDIO_WEB) {
                     mListener.audioFinished();
@@ -859,7 +874,7 @@ public class PlayerFragment extends BaseFragment implements
             case ExoPlayer.STATE_READY:
                 mediaSession.setActive(true);
 
-                if (isNeedToSeekToLatestListenPosition) {
+                if (isNeedToSeekToLatestListenPosition && !TextUtils.isEmpty(fileId)) {
                     long playerPosition = 0;
                     if (contentType != TYPE_AUDIO_LIVE && contentType != TYPE_VIDEO_LIVE) {
                         playerPosition = DataHelper.getPlayTime(getActivity().getContentResolver(), fileId);
@@ -905,7 +920,7 @@ public class PlayerFragment extends BaseFragment implements
     @Override
     public void onDestroyView() {
         super.onDestroyView();
-        if (deleteFileBeforeExit) {
+        if (deleteFileBeforeExit && !TextUtils.isEmpty(fileId)) {
             Logger.d("ExoPlayer.STATE_ENDED remove download content");
             if (contentType == TYPE_VIDEO_LOCAL) {
                 FileUtils.deleteVideoFile(fileId, getActivity());
@@ -1140,7 +1155,7 @@ public class PlayerFragment extends BaseFragment implements
 
     public void showNotification(boolean isLive, int mediaType) {
         Logger.d("showNotification()");
-        if (player == null) {
+        if (player == null || TextUtils.isEmpty(fileId)) {
             return;
         }
